requirements.txt
requests
psycopg2
spotipy
Pillow
git+git://github.com/simon987/hexlib.git
